+++ src/PH/Bundle/CoreBundle/Resources/config/doctrine/model/OrderItem.orm.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
+                  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
+                                      http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
+
+    <mapped-superclass name="PH\Component\Core\Model\OrderItem" table="ph_order_item">
+        <many-to-one field="subscription" target-entity="PH\Component\Subscription\Model\SubscriptionInterface">
+            <join-column name="subscription_id" referenced-column-name="id" />
+            <cascade>
+                <cascade-persist/>
+            </cascade>
+        </many-to-one>
+
+    </mapped-superclass>
+
+</doctrine-mapping>
