+++ src/PH/Bundle/CoreBundle/Resources/config/serializer/Model.Order.yml
@@ -1,10 +1,6 @@
 PH\Component\Core\Model\Order:
     exclusion_policy: ALL
     xml_root_name: order
-    properties:
-        subscription:
-            expose: true
-            groups: [Default, Detailed]
     relations:
         -   rel: self
             href:
