+++ src/PH/Bundle/SubscriptionBundle/Resources/config/doctrine/model/Subscription.orm.xml
@@ -13,6 +13,8 @@
         <field name="currencyCode" column="currency_code" length="3" />
         <field type="integer" name="amount" column="amount" />
         <field type="string" name="interval" column="interval" />
+        <field type="string" name="code" column="code" />
+        <field type="string" name="name" column="name" />
         <field name="createdAt" column="created_at" type="datetime">
             <gedmo:timestampable on="create"/>
         </field>
