+++ src/PH/Bundle/SubscriptionBundle/Resources/config/serializer/Model.Subscription.yml
@@ -17,6 +17,14 @@ PH\Component\Subscription\Model\Subscription:
             expose: true
             type: string
             groups: [Default, Detailed]
+        name:
+            expose: true
+            type: string
+            groups: [Default, Detailed]
+        code:
+            expose: true
+            type: string
+            groups: [Default, Detailed]
         createdAt:
             expose: true
             type: DateTime
