+++ src/PH/Component/Core/Model/Order.php
@@ -8,7 +8,6 @@ use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use PH\Component\Core\OrderCheckoutStates;
 use PH\Component\Core\OrderPaymentStates;
-use PH\Component\Subscription\Model\SubscriptionInterface;
 use Sylius\Component\Order\Model\Order as BaseOrder;
 use Webmozart\Assert\Assert;
 
@@ -43,11 +42,6 @@ class Order extends BaseOrder implements OrderInterface
      * @var string
      */
     protected $currencyCode;
-
-    /**
-     * @var SubscriptionInterface
-     */
-    protected $subscription;
 
     /**
      * Order constructor.
@@ -197,21 +191,5 @@ class Order extends BaseOrder implements OrderInterface
         Assert::string($currencyCode);
 
         $this->currencyCode = $currencyCode;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getSubscription(): SubscriptionInterface
-    {
-        return $this->subscription;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setSubscription(SubscriptionInterface $subscription): void
-    {
-        $this->subscription = $subscription;
     }
 }
