+++ src/PH/Component/Core/Model/OrderInterface.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace PH\Component\Core\Model;
 
 use Doctrine\Common\Collections\Collection;
-use PH\Component\Subscription\Model\SubscriptionInterface;
 use Sylius\Component\Order\Model\OrderInterface as BaseOrderInterface;
 
 interface OrderInterface extends BaseOrderInterface
@@ -88,14 +87,4 @@ interface OrderInterface extends BaseOrderInterface
      * @param string $currencyCode
      */
     public function setCurrencyCode(string $currencyCode): void;
-
-    /**
-     * @return SubscriptionInterface
-     */
-    public function getSubscription(): SubscriptionInterface;
-
-    /**
-     * @param SubscriptionInterface $subscription
-     */
-    public function setSubscription(SubscriptionInterface $subscription): void;
 }
