+++ src/PH/Component/Subscription/Model/Subscription.php
@@ -31,6 +31,16 @@ class Subscription implements SubscriptionInterface
     protected $interval = SubscriptionInterface::INTERVAL_MONTH;
 
     /**
+     * @var string
+     */
+    protected $name;
+
+    /**
+     * @var string
+     */
+    protected $code;
+
+    /**
      * Subscription constructor.
      */
     public function __construct()
@@ -92,5 +102,37 @@ class Subscription implements SubscriptionInterface
     public function setInterval(string $interval)
     {
         $this->interval = $interval;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getCode(): ?string
+    {
+        return $this->code;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setCode(string $code): void
+    {
+        $this->code = $code;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getName(): ?string
+    {
+        return $this->name;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setName(string $name): void
+    {
+        $this->name = $name;
     }
 }
