+++ src/PH/Component/Subscription/Model/SubscriptionInterface.php
@@ -47,4 +47,24 @@ interface SubscriptionInterface extends TimestampableInterface, ResourceInterfac
      * @param string $interval
      */
     public function setInterval(string $interval);
+
+    /**
+     * @return string|null
+     */
+    public function getCode(): ?string;
+
+    /**
+     * @param string $code
+     */
+    public function setCode(string $code): void;
+
+    /**
+     * @return string|null
+     */
+    public function getName(): ?string;
+
+    /**
+     * @param string $name
+     */
+    public function setName(string $name): void;
 }
